Create a Card-game(Procedural Programming)
This is a card game. In this game 8 cards will be selected from a pack of shuffled 52 cards. And then a card will be drawn out and then the user will be prompted to guess whether the next card will be high or lower than the previous card. If the user guesses right 20 points is added to their score and if they are wrong 15 points is deducted from their point. The starting point is 50 point for the user.